requirements.txt
asgiref==3.3.4
Django==3.2
django-activity-stream==0.10.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
djangorestframework==3.12.4
jsonfield==3.1.0
Markdown==3.3.4
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
